Split matchers out into separate .py files.  Add refresh_matchers, sync_matchers.
